+++ src/lib/index.ts
@@ -17,3 +17,4 @@
 export * from "./Common";
 export * from "./ServerSender";
 export * from "./Errors";
+export * from "./HeaderBuilder";