+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Added DKIM supports
 - Added dynamic domains management.
 - Fixed: the BASE64 output should not be more than 78 charactors each line.
+- Added tool class HeaderBuilder.
 
 ## v0.1.1